# [mysql] use compression protocol