--- a/xbmc/dbwrappers/dataset.cpp
+++ b/xbmc/dbwrappers/dataset.cpp
@@ -55,8 +55,9 @@ Database::~Database() {
   disconnect();		// Disconnect if connected to database
 }
 
-int Database::connectFull(const char *newHost, const char *newPort, const char *newDb, const char *newLogin, const char *newPasswd,
-                        const char *newKey, const char *newCert, const char *newCA, const char *newCApath, const char *newCiphers) {
+int Database::connectFull(const char *newHost, const char *newPort, const char *newDb, const char *newLogin,
+                          const char *newPasswd, const char *newKey, const char *newCert, const char *newCA,
+                          const char *newCApath, const char *newCiphers, bool newCompression) {
   host = newHost;
   port = newPort;
   db = newDb;
@@ -67,6 +68,7 @@ int Database::connectFull(const char *newHost, const char *newPort, const char *
   ca = newCA;
   capath = newCApath;
   ciphers = newCiphers;
+  compression = newCompression;
   return connect(true);
 }