--- a/xbmc/dbwrappers/dataset.h
+++ b/xbmc/dbwrappers/dataset.h
@@ -68,6 +68,7 @@ class Dataset;		// forward declaration of class Dataset
 class Database  {
 protected:
   bool active;
+  bool compression;
   std::string error, // Error description
     host, port, db, login, passwd, //Login info
     sequence_table, //Sequence table for nextid
@@ -108,13 +109,14 @@ public:
   const char *getSequenceTable(void) { return sequence_table.c_str(); }
 /* Get the default character set */
   const char *getDefaultCharset(void) { return default_charset.c_str(); }
-/* Sets SSL configuration */
-  virtual void setSSLConfig(const char *newKey, const char *newCert, const char *newCA, const char *newCApath, const char *newCiphers) {
+/* Sets configuration */
+  virtual void setConfig(const char *newKey, const char *newCert, const char *newCA, const char *newCApath, const char *newCiphers, bool newCompression) {
     key = newKey;
     cert = newCert;
     ca = newCA;
     capath = newCApath;
     ciphers = newCiphers;
+    compression = newCompression;
   }
 
 /* virtual methods that must be overloaded in derived classes */
@@ -128,7 +130,7 @@ public:
   virtual int connectFull( const char *newDb, const char *newHost=NULL,
                       const char *newLogin=NULL, const char *newPasswd=NULL,const char *newPort=NULL,
                       const char *newKey=NULL, const char *newCert=NULL, const char *newCA=NULL, 
-                      const char *newCApath=NULL, const char *newCiphers=NULL);
+                      const char *newCApath=NULL, const char *newCiphers=NULL, bool newCompression = false);
   virtual void disconnect(void) { active = false; }
   virtual int reset(void) { return DB_COMMAND_OK; }
   virtual int create(void) { return DB_COMMAND_OK; }