This enables the mysql compression protocol. The performance benefit is going to be largely dependent on the size of the result sets.