@popcornmix 2015-02-18
Any benchmark numbers? Espicially with a slow client or server.
A quick google showed someone warning this option is harmful due to high cpu:
http://karlssonondatabases.blogspot.co.uk/2012/08/mysql-cluster-performance-up-again-or.html