@mkortstiege 2015-02-18
No real benchmarks yet. Entering the movies node (`select * from movie_view`) was ~50% faster over a really slow internet connection.

Before:
```
DEBUG: RunQuery took 1088 ms for 359 items query: select * from movie_view
```
After:
```
EBUG: RunQuery took 465 ms for 359 items query: select * from movie_view
```
Did not notice any high cpu issues while testing on regular hardware. 

@MilhouseVH is testing this right now as well.