@MilhouseVH 2015-02-18
These are my results on three clients - a Pi1, Pi2, and Ubuntu x86.

In all cases the database is MySQL 5.5 on an HP N36L (8GB RAM, FreeNAS/FreeBSD), and all clients are wired into a GigE network.

"BEFORE" is without compression, "AFTER" is with compression.


##### Pi1 (1000MHz ARM, OpenELEC build #0217):

```
Movies (Titles):

BFFORE: 19:35:15 80369.109375 T:1968140288   DEBUG: RunQuery took 123 ms for 694 items query: select * from movie_view
BEFORE: 19:35:40 80394.382812 T:1968140288   DEBUG: RunQuery took 149 ms for 694 items query: select * from movie_view
BEFORE: 19:36:00 80414.015625 T:1968140288   DEBUG: RunQuery took 167 ms for 694 items query: select * from movie_view

AFTER : 19:43:11   286.509735 T:1968152576   DEBUG: RunQuery took 228 ms for 694 items query: select * from movie_view
AFTER : 19:43:29   304.470673 T:1968152576   DEBUG: RunQuery took 336 ms for 694 items query: select * from movie_view
AFTER : <file cached>

Movie Sets:

BEFORE: 19:30:42 80096.328125 T:1968033792   DEBUG: RunQuery took 89 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
BEFORE: 19:31:26 80140.125000 T:1968033792   DEBUG: RunQuery took 112 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
BEFORE: 19:31:48 80161.882812 T:1968033792   DEBUG: RunQuery took 93 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet

AFTER : 19:44:17   351.574890 T:1968152576   DEBUG: RunQuery took 113 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
AFTER : 19:44:35   370.232147 T:1968152576   DEBUG: RunQuery took 146 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
AFTER : 19:44:49   383.807373 T:1968152576   DEBUG: RunQuery took 174 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet

TV Shows:

BEFORE: 19:32:41 80215.062500 T:1968033792   DEBUG: RunQuery took 96 ms for 67 items query: SELECT * FROM tvshow_view
BEFORE: 19:32:52 80226.382812 T:1968033792   DEBUG: RunQuery took 72 ms for 67 items query: SELECT * FROM tvshow_view
BEFORE: 19:33:07 80240.671875 T:1968033792   DEBUG: RunQuery took 56 ms for 67 items query: SELECT * FROM tvshow_view

AFTER : 19:45:05   400.033478 T:1968152576   DEBUG: RunQuery took 120 ms for 67 items query: SELECT * FROM tvshow_view
AFTER : 19:45:25   420.437653 T:1968152576   DEBUG: RunQuery took 178 ms for 67 items query: SELECT * FROM tvshow_view
AFTER : 19:45:43   437.731140 T:1968152576   DEBUG: RunQuery took 157 ms for 67 items query: SELECT * FROM tvshow_view
```

In the case of the Pi1, it would more frequently use a local file cache when compression is enabled - Kodi will use a file cache if the query takes longer than a minimum amount of time, I believe. I tried to capture the query time several times, but Kodi would all too frequently switch to a file cache on the Pi1 with compression enabled.

##### Pi2 (1000Mhz ARM, OpenELEC build #0217):

```
Movies (Titles):

BFFORE: 18:54:56 77918.023438 T:1969049600   DEBUG: RunQuery took 104 ms for 694 items query: select * from movie_view
BEFORE: 19:03:02 78403.406250 T:1969049600   DEBUG: RunQuery took 103 ms for 694 items query: select * from movie_view
BEFORE: 19:03:10 78411.601562 T:1969049600   DEBUG: RunQuery took 103 ms for 694 items query: select * from movie_view

AFTER : 19:15:30   489.017029 T:1969033216   DEBUG: RunQuery took 170 ms for 694 items query: select * from movie_view
AFTER : 19:16:33   551.510742 T:1969033216   DEBUG: RunQuery took 283 ms for 694 items query: select * from movie_view
AFTER : 19:16:56   574.070129 T:1969033216   DEBUG: RunQuery took 173 ms for 694 items query: select * from movie_view

Movie Sets:

BEFORE: 19:03:21 78422.609375 T:1969049600   DEBUG: RunQuery took 60 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
BEFORE: 19:03:36 78437.742188 T:1969049600   DEBUG: RunQuery took 58 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet

AFTER : 19:17:13   591.907715 T:1969033216   DEBUG: RunQuery took 135 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
AFTER : 19:17:20   598.244263 T:1969033216   DEBUG: RunQuery took 112 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet

TV Shows:

BEFORE: 19:02:24 78365.750000 T:1969049600   DEBUG: RunQuery took 73 ms for 67 items query: SELECT * FROM tvshow_view
BEFORE: 19:02:38 78379.734375 T:1969049600   DEBUG: RunQuery took 71 ms for 67 items query: SELECT * FROM tvshow_view
BEFORE: 19:02:49 78390.679688 T:1969049600   DEBUG: RunQuery took 62 ms for 67 items query: SELECT * FROM tvshow_view

AFTER : 19:18:08   646.145569 T:1969033216   DEBUG: RunQuery took 92 ms for 67 items query: SELECT * FROM tvshow_view
AFTER : 19:18:16   654.265625 T:1969033216   DEBUG: RunQuery took 113 ms for 67 items query: SELECT * FROM tvshow_view
AFTER : 19:18:23   661.145630 T:1969033216   DEBUG: RunQuery took 111 ms for 67 items query: SELECT * FROM tvshow_view
```

Much less reliance on the file cache while using the Pi2, with or without compression.

##### x86 (AMD FX-8350 4GHz CPU, Ubuntu 14.10, Kodi master):

```
Movies (Titles):

BFFORE: 19:56:26 T:140267865766080   DEBUG: RunQuery took 104 ms for 694 items query: select * from movie_view
BEFORE: 19:56:37 T:140267865766080   DEBUG: RunQuery took 13 ms for 694 items query: select * from movie_view
BEFORE: 19:56:45 T:140267865766080   DEBUG: RunQuery took 53 ms for 694 items query: select * from movie_view

AFTER : 20:02:42 T:140089844930752   DEBUG: RunQuery took 123 ms for 694 items query: select * from movie_view
AFTER : 20:02:47 T:140089844930752   DEBUG: RunQuery took 121 ms for 694 items query: select * from movie_view
AFTER : 20:02:52 T:140089844930752   DEBUG: RunQuery took 122 ms for 694 items query: select * from movie_view

Movie Sets:

BEFORE: 19:58:21 T:140267865766080   DEBUG: RunQuery took 23 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
BEFORE: 19:58:31 T:140267865766080   DEBUG: RunQuery took 10 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
BEFORE: 19:58:34 T:140267865766080   DEBUG: RunQuery took 46 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet

AFTER : 20:03:45 T:140089844930752   DEBUG: RunQuery took 49 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
AFTER : 20:03:50 T:140089844930752   DEBUG: RunQuery took 48 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet
AFTER : 20:03:54 T:140089844930752   DEBUG: RunQuery took 66 ms for 269 items query: select * from movie_view  JOIN sets ON movie_view.idSet = sets.idSet ORDER BY sets.idSet

TV Shows:

BEFORE: 19:59:09 T:140267865766080   DEBUG: RunQuery took 2 ms for 67 items query: SELECT * FROM tvshow_view
BEFORE: 19:59:28 T:140267865766080   DEBUG: RunQuery took 3 ms for 67 items query: SELECT * FROM tvshow_view
BEFORE: 19:59:32 T:140267865766080   DEBUG: RunQuery took 3 ms for 67 items query: SELECT * FROM tvshow_view

AFTER : 20:04:30 T:140089844930752   DEBUG: RunQuery took 7 ms for 67 items query: SELECT * FROM tvshow_view
AFTER : 20:04:36 T:140089844930752   DEBUG: RunQuery took 7 ms for 67 items query: SELECT * FROM tvshow_view
AFTER : 20:04:40 T:140089844930752   DEBUG: RunQuery took 13 ms for 67 items query: SELECT * FROM tvshow_view
```

In every case, the queries are slower when compression is enabled, usually by 50% or more, even on the Ubuntu client. This would suggest the problem (extra overhead) is with the HP N36L server, and not the clients. However the server shows very little load during the queries (2-3% max) but given the short duration of the queries the actual load could be 100% for only 100ms or so. Either way, it appears the AMD Athlon II Neo 1.3GHz CPU isn't too hot on zlib compression.

If the MySQL compression feature is to be supported, I'd definitely suggest an advancedsetting so that it can be enabled as it should be left disabled by default. If my results are typical of low-end server hardware, then compression may only be of benefit when the MySQL server is reasonably powerful (i3+), which probably rules out a lot of server hardware that is currently used to run MySQL for Kodi (Synology etc.).

Even with a low-end server, compression might still show some benefit over a WiFi (or internet) connection. Perhaps someone can capture query timings with/without compression over a WiFi connection.