@MilhouseVH 2015-02-18
No, I don't think that information is available in the Kodi log, and it's probably transparent to the client.  Observing network traffic isn't reliable either, as it can include NFS traffic. I guess sniffing the wire might be one solution.

This is the difference in wire traffic RX/TX when entering Movies (Titles) on the Pi2:

##### Without compresson:
```
Time         ARM    Core    H264 Core Temp (Max)  IRQ/s     RX B/s     TX B/s GPUMem Free  %user  %nice   %sys  %idle  %iowt   %irq %s/irq %total   cpu0   cpu1   cpu2   cpu3 Memory Free/Used
======== ======= ======= ======= =============== ====== ========== ========== =========== ====== ====== ====== ====== ====== ====== ====== ====== ====== ====== ====== ====== ================

21:04:08 1000Mhz  500Mhz    0Mhz 49.23C (50.31C)  2,006  1,286,230     95,191 239M ( 79%)  19.49   0.93   3.71  73.08   0.00   0.00   0.46  26.92  50.82  12.77  27.62  16.48 625,260 kB/ 9.5%
21:04:09 1000Mhz  500Mhz    0Mhz 50.84C (50.84C)  3,377  1,053,228    214,292 238M ( 79%)  11.41   0.70   4.43  80.12   0.00   0.00   0.70  19.88  44.10   4.04   8.70  23.61 625,288 kB/ 9.5%
21:04:11 1000Mhz  500Mhz    0Mhz 48.15C (50.84C)  1,407    289,356     63,519 238M ( 79%)   4.02   0.47   1.66  92.57   0.00   0.00   0.00   7.43  13.82   3.40   1.51  10.98 625,296 kB/ 9.5%
21:04:12 1000Mhz  500Mhz    0Mhz 48.69C (50.84C)    630        233        472 238M ( 79%)   1.42   0.71   0.95  97.00   0.00   0.00   0.00   3.00   1.58   2.53   0.63   6.31 625,648 kB/ 9.4%
...
21:04:21 1000Mhz  500Mhz    0Mhz 50.31C (50.84C)  1,978  1,273,964     92,904 246M ( 82%)  17.94   0.47   4.19  73.62   0.00   0.00   0.93  26.38  73.91  10.53  13.33   7.74 624,400 kB/ 9.6%
21:04:22 1000Mhz  500Mhz    0Mhz 50.84C (50.84C)  3,343  1,043,820    211,544 238M ( 79%)  10.50   0.93   4.20  80.00   0.00   0.00   0.93  20.00  41.23  18.84  11.37   8.57 624,040 kB/ 9.7%
21:04:23  999Mhz  500Mhz    0Mhz 49.23C (50.84C)  1,483    316,992     69,049 238M ( 79%)   3.56   0.24   2.13  92.49   0.00   0.00   0.47   7.51  15.58  10.83   3.24   1.35 624,292 kB/ 9.6%
21:04:24 1000Mhz  500Mhz    0Mhz 48.15C (50.84C)    630        167        451 238M ( 79%)   1.42   0.47   0.71  96.84   0.00   0.00   0.00   3.16   3.16   5.06   3.16   1.27 624,328 kB/ 9.6%
21:04:25 1000Mhz  500Mhz    0Mhz 50.31C (50.84C)  1,220    174,102     42,977 238M ( 79%)   5.85   2.58   3.75  86.39   0.00   0.00   0.00  13.61  13.84  27.89   6.35   5.41 624,064 kB/ 9.7%
...
21:04:31 1000Mhz  500Mhz    0Mhz 49.77C (50.84C)  1,937  1,237,659     90,323 246M ( 82%)  17.87   0.46   3.90  74.69   0.00   0.00   0.92  25.31  72.51  15.68   8.35   4.69 623,368 kB/ 9.8%
21:04:32 1000Mhz  500Mhz    0Mhz 51.38C (51.38C)  3,336  1,041,987    210,244 238M ( 79%)  12.43   1.17   4.22  79.76   0.00   0.00   0.94  20.24  46.51   5.22  23.99   4.28 623,728 kB/ 9.7%
21:04:33 1000Mhz  500Mhz    0Mhz 49.77C (51.38C)  1,546    341,506     74,162 238M ( 79%)   3.32   0.71   1.90  92.01   0.00   0.00   0.24   7.99  15.58   4.20  10.84   2.30 623,556 kB/ 9.8%
21:04:34 1000Mhz  500Mhz    0Mhz 48.69C (51.38C)    630         87        474 238M ( 79%)   1.66   0.47   0.95  96.77   0.00   0.00   0.00   3.23   1.34   3.23   6.08   1.34 623,700 kB/ 9.7%
21:04:35  999Mhz  500Mhz    0Mhz 51.38C (51.38C)  1,228    174,876     43,608 238M ( 79%)   7.52   2.58   2.82  85.05   0.00   0.00   0.47  14.95  22.00  28.58   6.97   1.33 623,280 kB/ 9.8%^
```
##### With compression:

```
Time         ARM    Core    H264 Core Temp (Max)  IRQ/s     RX B/s     TX B/s GPUMem Free  %user  %nice   %sys  %idle  %iowt   %irq %s/irq %total   cpu0   cpu1   cpu2   cpu3 Memory Free/Used
======== ======= ======= ======= =============== ====== ========== ========== =========== ====== ====== ====== ====== ====== ====== ====== ====== ====== ====== ====== ====== ================

20:58:55 1000Mhz  500Mhz    0Mhz 46.54C (47.62C)  1,432    423,714     43,497 232M ( 77%)  15.88   0.92   2.99  77.33   0.00   0.00   0.69  22.67  39.24  33.71  10.70   7.02 616,740 kB/10.7%
20:58:56 1000Mhz  500Mhz    0Mhz 49.77C (49.77C)  2,524    292,038    136,264 208M ( 69%)  15.48   1.39   7.16  73.25   0.00   0.00   0.92  26.75  58.41  20.51  14.97  13.12 617,568 kB/10.6%
20:58:58 1000Mhz  500Mhz    0Mhz 49.23C (49.77C)  2,646    282,784    159,993 208M ( 69%)  12.91   0.94   3.75  80.96   0.00   0.00   0.70  19.04  54.94  14.58   4.25   2.37 617,420 kB/10.6%
20:58:59 1000Mhz  500Mhz    0Mhz 47.62C (49.77C)    788      3,910      2,826 208M ( 69%)   1.43   0.48   0.95  96.74   0.00   0.00   0.00   3.26   3.02   6.82   3.02   1.12 617,576 kB/10.6%
...
20:59:13 1000Mhz  500Mhz    0Mhz 47.08C (49.77C)  1,034    368,180     15,246 240M ( 80%)  13.48   0.47   1.42  83.49   0.00   0.00   0.00  16.51   5.40  53.64   3.50   2.56 617,328 kB/10.7%
20:59:14 1000Mhz  500Mhz    0Mhz 49.77C (49.77C)  2,474    288,387    126,167 208M ( 69%)  17.76   1.38   7.84  70.11   0.00   0.00   0.46  29.89  57.56  39.11  19.74   4.05 617,708 kB/10.6%
20:59:16 1000Mhz  500Mhz    0Mhz 49.77C (49.77C)  2,599    275,538    155,900 208M ( 69%)  13.63   0.94   3.52  79.88   0.00   0.00   1.17  20.12  49.25  18.24   6.96   6.02 617,396 kB/10.6%
20:59:17 1000Mhz  500Mhz    0Mhz 48.69C (49.77C)  1,331     84,431     49,221 208M ( 69%)   6.13   0.94   2.12  89.84   0.47   0.00   0.24  10.16  18.88  15.11   0.96   4.74 617,336 kB/10.7%
...
20:59:28 1000Mhz  500Mhz    0Mhz 48.69C (50.84C)    963    276,188     10,824 234M ( 78%)   3.79   0.71   1.66  92.99   0.00   0.00   0.47   7.01  20.50   2.52   3.47   1.57 617,204 kB/10.7%
20:59:29  999Mhz  500Mhz    0Mhz 49.23C (50.84C)  1,806    257,121     75,799 234M ( 78%)  20.26   0.47   3.30  73.75   0.00   0.00   0.94  26.25  73.61  25.54   3.86   1.98 617,036 kB/10.7%
20:59:30 1000Mhz  500Mhz    0Mhz 51.38C (51.38C)  2,626    292,911    150,049 226M ( 75%)  14.04   0.70   4.21  77.69   0.00   0.00   0.94  22.31  55.07  25.12   5.47   3.60 617,044 kB/10.7%
20:59:31 1000Mhz  500Mhz    0Mhz 50.31C (51.38C)  2,067    194,893    111,322 226M ( 75%)   9.72   0.71   2.37  85.54   0.00   0.00   0.24  14.46  36.49  15.64   4.27   0.47 617,292 kB/10.7%
```

So there is a significant reduction in the RX figures, as much as 75% reduction, although I'm not able to say how much of this is SQL related (though given that the only difference between the tests is the addition of compression, perhaps it's entirely SQL data, though not sure why the transfers are multi-second when the queries are completing sub-second).

Benchmarking compression properly probably requires a small test client (C++, same libraries as Kodi) that can be used to compare the benefits of MySQL compression for a variety of typical Kodi queries. This would yield a much greater degree of accuracy/confidence than you will get from analysing the Kodi debug log, which is at best indicative.

I'll include this PR in tonights Pi1/Pi2 OpenELEC test builds, which may allow us to capture more results on different networks and server setups.