@mkortstiege 2015-02-19
Here are my timings using a Core i7-4770 @ 3.40GHz running Ubuntu 14.10 as client and a Pentium 4 @ 2.40GHz running the database (MySQL 5.5) on a GigE network.

##### Without compresson:
```
DEBUG: RunQuery took 295 ms for 500 items query: select * from movie_view
DEBUG: RunQuery took 287 ms for 500 items query: select * from movie_view
DEBUG: RunQuery took 283 ms for 500 items query: select * from movie_view
```

##### With compresson:
```
DEBUG: RunQuery took 231 ms for 500 items query: select * from movie_view
DEBUG: RunQuery took 206 ms for 500 items query: select * from movie_view
DEBUG: RunQuery took 231 ms for 500 items query: select * from movie_view
```

Apparently the performance gain is pretty low compared to using the same server over the internet. So the real benefit using compression is dependent on network bandwidth, latency between the database and the clients and the the actual client that has to handle the zlib compression.